stream/twitter_to_redis/requirements.txt
python-twitter
requests
redis
flask
